Rename datastore controller to keyvalue controller
